Fix additional MSVC warnings in Decompress, MyNetCDF, dbd2netCDF
- Decompress.C: Cast size_t to int for LZ4_decompress_safe argument
- MyNetCDF.C: Cast nDims to int for nc_def_var, use std::nanf for float
- dbd2netCDF.C: Rename shadowed loop variable i to j, cast size_t to unsigned int

+++ src/Decompress.C
@@ -38,7 +38,7 @@ int DecompressTWRBuf::underflow() {
     if (!this->mIS.read(frame.data(), n)) { // EOF
       return std::char_traits<char>::eof();
     }
-    const size_t j(LZ4_decompress_safe(frame.data(), this->mBuffer, n, sizeof(this->mBuffer)));
+    const size_t j(LZ4_decompress_safe(frame.data(), this->mBuffer, static_cast<int>(n), sizeof(this->mBuffer)));
     if (j > sizeof(this->mBuffer)) { // Probably a corrupted file
       LOG_ERROR("Attempt to decompress lz4 block with too much data: {} > {} in {} (block size {})",
                 j, sizeof(this->mBuffer), this->mFilename, n);
